change cell metadata actions to cellOperations.
